+++ components/pm-subscription-handler/Changelog.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [2.2.0]
 ### Changed
 * Update Filter API (DCAEGEN2-2922)
+* Create Measurement Group API (DCAEGEN2-2920)
+
 
 ## [2.1.1]
 ### Changed